+++ src/commands/webhooks.ts
@@ -1,33 +1,18 @@
 import * as vscode from 'vscode';
 import { IContext, showErrorMessage } from '../common';
 import { IWebhook, IWebhookEvent } from '../providers/webhooks';
-import { WebhookSystem } from 'aps-sdk-node';
 import { withProgress, createWebViewPanel } from '../common';
-
-function webhookScopes(system: WebhookSystem): string[] {
-	let scopes: string[] = [];
-	switch (system) {
-		case WebhookSystem.Data:
-		case WebhookSystem.RevitCloudWorksharing:
-			scopes.push('folder');
-			break;
-		case WebhookSystem.Derivative:
-			scopes.push('workflow');
-			break;
-		case WebhookSystem.FusionLifecycle:
-			scopes.push('workspace');
-			scopes.push('workflow.transition');
-			break;
-	}
-	return scopes;
-}
+import { WebhookEvent, WEBHOOKS, WebhookSystem } from '../interfaces/webhooks';
 
 export async function createWebhook({ system, event }: IWebhookEvent, context: IContext, successCallback?: () => void) {
-	const scopes = webhookScopes(system);
+	const _system = WEBHOOKS.find(webhook => webhook.id === system) as WebhookSystem;
+	const _event = _system.events.find(ev => ev.id === event) as WebhookEvent;
+	const { scopes } = _event;
 	let panel = createWebViewPanel(context, 'create-webhook.js', 'create-webhook', 'Create Webhook', { system, event, scopes }, async message => {
 		switch (message.command) {
 			case 'create':
 				try {
+					// @ts-ignore
 					const id = await withProgress(`Creating webhook`, context.webhookClient.createHook(system, event, {
 						callbackUrl: message.webhook.callbackUrl,
 						// @ts-ignore
@@ -50,6 +35,7 @@ export async function createWebhook({ system, event }: IWebhookEvent, context: I
 
 export async function viewWebhookDetails({ id, system, event }: IWebhook, context: IContext) {
 	try {
+		// @ts-ignore
 		const webhookDetail = await withProgress(`Getting webhook details: ${id}`, context.webhookClient.getHookDetails(system, event, id));
 		createWebViewPanel(context, 'webhook-details.js', 'webhook-details', `Webhook Details: ${id}`, { detail: webhookDetail });
 	} catch(err) {
@@ -59,6 +45,7 @@ export async function viewWebhookDetails({ id, system, event }: IWebhook, contex
 
 export async function deleteWebhook({ system, event, id }: IWebhook, context: IContext) {
 	try {
+		// @ts-ignore
 		await withProgress(`Removing webhook: ${id}`, context.webhookClient.deleteHook(system, event, id));
 		vscode.window.showInformationMessage(`Webhook removed: ${id}`);
 	} catch(err) {
@@ -68,12 +55,16 @@ export async function deleteWebhook({ system, event, id }: IWebhook, context: IC
 
 export async function updateWebhook({ system, event, id }: IWebhook, context: IContext, successCallback?: () => void) {
 	try {
+		// @ts-ignore
 		const webhookDetail = await withProgress(`Retrieving webhook data: ${id}`, context.webhookClient.getHookDetails(system, event, id));
-		const scopes = webhookScopes(system);
+		const _system = WEBHOOKS.find(webhook => webhook.id === system) as WebhookSystem;
+		const _event = _system.events.find(ev => ev.id === event) as WebhookEvent;
+		const { scopes } = _event;
 		let panel = createWebViewPanel(context, 'update-webhook.js', 'update-webhook', `Update Webhook: ${id}`, { detail: webhookDetail, scopes }, async message => {
 			switch (message.command) {
 				case 'update':
 					try {
+						// @ts-ignore
 						await withProgress(`Updating webhook ${id}`, context.webhookClient.updateHook(system, event, id, {
 							// status: message.webhook.status,
 							filter: message.webhook.filter || null,